Fixed some spelling errors in main.cpp and let the formatter do his thing in the states files (No real changes in the content)

+++ CMP3060 Phase 1 - Student Version/source/states/clear_color_state.hpp
@@ -2,31 +2,34 @@
 
 #include <application.hpp>
 
-class ClearColorState: public our::State {
-    private:
-    double r = 0.0;
-    double g = 0.0;
-    double b = 0.0;
-    double a = 0.0;
-    
+class ClearColorState : public our::State
+{
+private:
+    GLfloat r = 0.0;
+    GLfloat g = 0.0;
+    GLfloat b = 0.0;
+    GLfloat a = 0.0;
+
     // onInitialize() function is called once before the state starts
-    void onInitialize() override {
-        //TODO: Read the color from the configuration file and use it to set the clear color for the window
-        //HINT: you can the configuration for the whole application using "getApp()->getConfig()"
-        //To see how the clear color is written in the json files, see "config/blue-screen.json"
-        //To know how read data from a nlohmann::json object, 
-        //look at the following documentation: https://json.nlohmann.me/features/element_access/
-        const nlohmann::json& clearColor = getApp()->getConfig().at("scene").at("clear-color");
+    void onInitialize() override
+    {
+        // TODO: Read the color from the configuration file and use it to set the clear color for the window
+        // HINT: you can the configuration for the whole application using "getApp()->getConfig()"
+        // To see how the clear color is written in the json files, see "config/blue-screen.json"
+        // To know how read data from a nlohmann::json object,
+        // look at the following documentation: https://json.nlohmann.me/features/element_access/
+        const nlohmann::json &clearColor = getApp()->getConfig().at("scene").at("clear-color");
         r = clearColor.at("r");
         g = clearColor.at("g");
         b = clearColor.at("b");
         a = clearColor.at("a");
     }
 
-    // onDraw(deltaTime) function is called every frame 
-    void onDraw(double deltaTime) override {
-        //At the start of frame we want to clear the screen. Otherwise we would still see the results from the previous frame.
-        glClearColor(r , g , b , a);
+    // onDraw(deltaTime) function is called every frame
+    void onDraw(double deltaTime) override
+    {
+        // At the start of frame we want to clear the screen. Otherwise we would still see the results from the previous frame.
+        glClearColor(r, g, b, a);
         glClear(GL_COLOR_BUFFER_BIT);
     }
 };
